requirements.txt
lxml=4.7.1
openpyxl=3.0.9
pandas=1.3.5
requests=2.26.0
svg.path=4.1
xlrd=2.0.1
